move appendable_mmap_dense_vector_storage
