+++ lib/segment/src/vector_storage/dense/mod.rs
@@ -1,2 +1,3 @@
 pub mod memmap_dense_vector_storage;
 pub mod mmap_dense_vectors;
+pub mod appendable_mmap_dense_vector_storage;
