+++ lib/segment/src/vector_storage/mod.rs
@@ -1,4 +1,3 @@
-pub mod appendable_mmap_dense_vector_storage;
 #[cfg(target_os = "linux")]
 pub mod async_raw_scorer;
 mod chunked_mmap_vectors;
